requirements.txt
discord.py
qrcode